requirements.txt
streamlit==1.41.0
#setuptools>=75.6.0
pandas>=2.2,<3.0
numpy>=2.1,<2.2
euroleague_api
Pillow>=11.0,<12.0
plotly
matplotlib
statsmodels>=0.14.2
plotly==5.24.1
#scikit-learn==1.2.0
